+++ pandas/tests/series/test_operators.py
@@ -2,7 +2,6 @@
 # pylint: disable-msg=E1101,W0612
 
 import pytest
-import pytz
 
 from collections import Iterable
 from datetime import datetime, timedelta
@@ -19,7 +18,6 @@ from pandas.core.indexes.datetimes import Timestamp
 from pandas.core.indexes.timedeltas import Timedelta
 import pandas.core.nanops as nanops
 
-from pandas.errors import PerformanceWarning
 from pandas.compat import range, zip
 from pandas import compat
 from pandas.util.testing import (assert_series_equal, assert_almost_equal,
@@ -30,19 +28,6 @@ from .common import TestData
 
 
 class TestSeriesComparisons(object):
-    def test_series_comparison_scalars(self):
-        series = Series(date_range('1/1/2000', periods=10))
-
-        val = datetime(2000, 1, 4)
-        result = series > val
-        expected = Series([x > val for x in series])
-        tm.assert_series_equal(result, expected)
-
-        val = series[5]
-        result = series > val
-        expected = Series([x > val for x in series])
-        tm.assert_series_equal(result, expected)
-
     def test_comparisons(self):
         left = np.random.randn(10)
         right = np.random.randn(10)
@@ -62,33 +47,6 @@ class TestSeriesComparisons(object):
         exp = Series([False, False, False])
         assert_series_equal(s == s2, exp)
         assert_series_equal(s2 == s, exp)
-
-    def test_operator_series_comparison_zerorank(self):
-        # GH 13006
-        result = np.float64(0) > pd.Series([1, 2, 3])
-        expected = 0.0 > pd.Series([1, 2, 3])
-        tm.assert_series_equal(result, expected)
-        result = pd.Series([1, 2, 3]) < np.float64(0)
-        expected = pd.Series([1, 2, 3]) < 0.0
-        tm.assert_series_equal(result, expected)
-        result = np.array([0, 1, 2])[0] > pd.Series([0, 1, 2])
-        expected = 0.0 > pd.Series([1, 2, 3])
-        tm.assert_series_equal(result, expected)
-
-    def test_object_comparisons(self):
-        s = Series(['a', 'b', np.nan, 'c', 'a'])
-
-        result = s == 'a'
-        expected = Series([True, False, False, False, True])
-        assert_series_equal(result, expected)
-
-        result = s < 'a'
-        expected = Series([False, False, False, False, False])
-        assert_series_equal(result, expected)
-
-        result = s != 'a'
-        expected = -(s == 'a')
-        assert_series_equal(result, expected)
 
     def test_categorical_comparisons(self):
         # GH 8938
@@ -284,27 +242,6 @@ class TestSeriesComparisons(object):
         tm.assert_series_equal(cat == "d", Series([False, False, False]))
         tm.assert_series_equal(cat != "d", Series([True, True, True]))
 
-    @pytest.mark.parametrize('dtype', [None, object])
-    def test_more_na_comparisons(self, dtype):
-        left = Series(['a', np.nan, 'c'], dtype=dtype)
-        right = Series(['a', np.nan, 'd'], dtype=dtype)
-
-        result = left == right
-        expected = Series([True, False, False])
-        assert_series_equal(result, expected)
-
-        result = left != right
-        expected = Series([False, True, True])
-        assert_series_equal(result, expected)
-
-        result = left == np.nan
-        expected = Series([False, False, False])
-        assert_series_equal(result, expected)
-
-        result = left != np.nan
-        expected = Series([True, True, True])
-        assert_series_equal(result, expected)
-
     @pytest.mark.parametrize('pair', [
         ([pd.Timestamp('2011-01-01'), NaT, pd.Timestamp('2011-01-03')],
          [NaT, NaT, pd.Timestamp('2011-01-03')]),
@@ -344,34 +281,6 @@ class TestSeriesComparisons(object):
 
         expected = Series([False, False, True])
         assert_series_equal(left <= right, expected)
-
-    @pytest.mark.parametrize('data', [
-        [pd.Timestamp('2011-01-01'), NaT, pd.Timestamp('2011-01-03')],
-        [pd.Timedelta('1 days'), NaT, pd.Timedelta('3 days')],
-        [pd.Period('2011-01', freq='M'), NaT, pd.Period('2011-03', freq='M')]
-    ])
-    @pytest.mark.parametrize('dtype', [None, object])
-    def test_nat_comparisons_scalar(self, dtype, data):
-        left = Series(data, dtype=dtype)
-
-        expected = Series([False, False, False])
-        assert_series_equal(left == pd.NaT, expected)
-        assert_series_equal(pd.NaT == left, expected)
-
-        expected = Series([True, True, True])
-        assert_series_equal(left != pd.NaT, expected)
-        assert_series_equal(pd.NaT != left, expected)
-
-        expected = Series([False, False, False])
-        assert_series_equal(left < pd.NaT, expected)
-        assert_series_equal(pd.NaT > left, expected)
-        assert_series_equal(left <= pd.NaT, expected)
-        assert_series_equal(pd.NaT >= left, expected)
-
-        assert_series_equal(left > pd.NaT, expected)
-        assert_series_equal(pd.NaT < left, expected)
-        assert_series_equal(left >= pd.NaT, expected)
-        assert_series_equal(pd.NaT <= left, expected)
 
     def test_comparison_different_length(self):
         a = Series(['a', 'b', 'c'])
@@ -591,73 +500,6 @@ class TestSeriesComparisons(object):
 
 class TestTimedeltaSeriesArithmetic(object):
 
-    def test_timedelta64_operations_with_DateOffset(self):
-        # GH 10699
-        td = Series([timedelta(minutes=5, seconds=3)] * 3)
-        result = td + pd.offsets.Minute(1)
-        expected = Series([timedelta(minutes=6, seconds=3)] * 3)
-        assert_series_equal(result, expected)
-
-        result = td - pd.offsets.Minute(1)
-        expected = Series([timedelta(minutes=4, seconds=3)] * 3)
-        assert_series_equal(result, expected)
-
-        with tm.assert_produces_warning(PerformanceWarning):
-            result = td + Series([pd.offsets.Minute(1), pd.offsets.Second(3),
-                                  pd.offsets.Hour(2)])
-        expected = Series([timedelta(minutes=6, seconds=3), timedelta(
-            minutes=5, seconds=6), timedelta(hours=2, minutes=5, seconds=3)])
-        assert_series_equal(result, expected)
-
-        result = td + pd.offsets.Minute(1) + pd.offsets.Second(12)
-        expected = Series([timedelta(minutes=6, seconds=15)] * 3)
-        assert_series_equal(result, expected)
-
-        # valid DateOffsets
-        for do in ['Hour', 'Minute', 'Second', 'Day', 'Micro', 'Milli',
-                   'Nano']:
-            op = getattr(pd.offsets, do)
-            td + op(5)
-            op(5) + td
-            td - op(5)
-            op(5) - td
-
-    def test_timedelta64_operations_with_timedeltas(self):
-        # td operate with td
-        td1 = Series([timedelta(minutes=5, seconds=3)] * 3)
-        td2 = timedelta(minutes=5, seconds=4)
-        result = td1 - td2
-        expected = (Series([timedelta(seconds=0)] * 3) -
-                    Series([timedelta(seconds=1)] * 3))
-        assert result.dtype == 'm8[ns]'
-        assert_series_equal(result, expected)
-
-        result2 = td2 - td1
-        expected = (Series([timedelta(seconds=1)] * 3) -
-                    Series([timedelta(seconds=0)] * 3))
-        assert_series_equal(result2, expected)
-
-        # roundtrip
-        assert_series_equal(result + td2, td1)
-
-        # Now again, using pd.to_timedelta, which should build
-        # a Series or a scalar, depending on input.
-        td1 = Series(pd.to_timedelta(['00:05:03'] * 3))
-        td2 = pd.to_timedelta('00:05:04')
-        result = td1 - td2
-        expected = (Series([timedelta(seconds=0)] * 3) -
-                    Series([timedelta(seconds=1)] * 3))
-        assert result.dtype == 'm8[ns]'
-        assert_series_equal(result, expected)
-
-        result2 = td2 - td1
-        expected = (Series([timedelta(seconds=1)] * 3) -
-                    Series([timedelta(seconds=0)] * 3))
-        assert_series_equal(result2, expected)
-
-        # roundtrip
-        assert_series_equal(result + td2, td1)
-
     def test_operators_timedelta64(self):
         # series ops
         v1 = date_range('2012-1-1', periods=3, freq='D')
@@ -807,40 +649,8 @@ class TestTimedeltaSeriesArithmetic(object):
         assert_series_equal(timedelta_series / nan,
                             nat_series_dtype_timedelta)
 
-    @pytest.mark.parametrize('scalar_td', [timedelta(minutes=5, seconds=4),
-                                           Timedelta(minutes=5, seconds=4),
-                                           Timedelta('5m4s').to_timedelta64()])
-    def test_operators_timedelta64_with_timedelta(self, scalar_td):
-        # smoke tests
-        td1 = Series([timedelta(minutes=5, seconds=3)] * 3)
-        td1.iloc[2] = np.nan
-
-        td1 + scalar_td
-        scalar_td + td1
-        td1 - scalar_td
-        scalar_td - td1
-        td1 / scalar_td
-        scalar_td / td1
-
 
 class TestDatetimeSeriesArithmetic(object):
-    @pytest.mark.parametrize(
-        'box, assert_func',
-        [(Series, tm.assert_series_equal),
-         (pd.Index, tm.assert_index_equal)])
-    def test_sub_datetime64_not_ns(self, box, assert_func):
-        # GH#7996
-        dt64 = np.datetime64('2013-01-01')
-        assert dt64.dtype == 'datetime64[D]'
-
-        obj = box(date_range('20130101', periods=3))
-        res = obj - dt64
-        expected = box([Timedelta(days=0), Timedelta(days=1),
-                        Timedelta(days=2)])
-        assert_func(res, expected)
-
-        res = dt64 - obj
-        assert_func(res, -expected)
 
     def test_operators_datetimelike_invalid(self, all_arithmetic_operators):
         # these are all TypeEror ops
@@ -912,283 +722,6 @@ class TestDatetimeSeriesArithmetic(object):
         # ## timetimedelta with datetime64 ###
         td1 + dt1
         dt1 + td1
-
-    def test_operators_datetimelike_with_timezones(self):
-
-        tz = 'US/Eastern'
-        dt1 = Series(date_range('2000-01-01 09:00:00', periods=5,
-                                tz=tz), name='foo')
-        dt2 = dt1.copy()
-        dt2.iloc[2] = np.nan
-
-        td1 = Series(timedelta_range('1 days 1 min', periods=5, freq='H'))
-        td2 = td1.copy()
-        td2.iloc[1] = np.nan
-
-        result = dt1 + td1[0]
-        exp = (dt1.dt.tz_localize(None) + td1[0]).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        result = dt2 + td2[0]
-        exp = (dt2.dt.tz_localize(None) + td2[0]).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        # odd numpy behavior with scalar timedeltas
-        result = td1[0] + dt1
-        exp = (dt1.dt.tz_localize(None) + td1[0]).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        result = td2[0] + dt2
-        exp = (dt2.dt.tz_localize(None) + td2[0]).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        result = dt1 - td1[0]
-        exp = (dt1.dt.tz_localize(None) - td1[0]).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-        with pytest.raises(TypeError):
-            td1[0] - dt1
-
-        result = dt2 - td2[0]
-        exp = (dt2.dt.tz_localize(None) - td2[0]).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-        with pytest.raises(TypeError):
-            td2[0] - dt2
-
-        result = dt1 + td1
-        exp = (dt1.dt.tz_localize(None) + td1).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        result = dt2 + td2
-        exp = (dt2.dt.tz_localize(None) + td2).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        result = dt1 - td1
-        exp = (dt1.dt.tz_localize(None) - td1).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        result = dt2 - td2
-        exp = (dt2.dt.tz_localize(None) - td2).dt.tz_localize(tz)
-        assert_series_equal(result, exp)
-
-        with pytest.raises(TypeError):
-            td1 - dt1
-        with pytest.raises(TypeError):
-            td2 - dt2
-
-    def test_sub_single_tz(self):
-        # GH12290
-        s1 = Series([pd.Timestamp('2016-02-10', tz='America/Sao_Paulo')])
-        s2 = Series([pd.Timestamp('2016-02-08', tz='America/Sao_Paulo')])
-        result = s1 - s2
-        expected = Series([Timedelta('2days')])
-        assert_series_equal(result, expected)
-        result = s2 - s1
-        expected = Series([Timedelta('-2days')])
-        assert_series_equal(result, expected)
-
-    def test_dt64tz_series_sub_dtitz(self):
-        # GH#19071 subtracting tzaware DatetimeIndex from tzaware Series
-        # (with same tz) raises, fixed by #19024
-        dti = pd.date_range('1999-09-30', periods=10, tz='US/Pacific')
-        ser = pd.Series(dti)
-        expected = pd.Series(pd.TimedeltaIndex(['0days'] * 10))
-
-        res = dti - ser
-        tm.assert_series_equal(res, expected)
-        res = ser - dti
-        tm.assert_series_equal(res, expected)
-
-    def test_sub_datetime_compat(self):
-        # see gh-14088
-        s = Series([datetime(2016, 8, 23, 12, tzinfo=pytz.utc), pd.NaT])
-        dt = datetime(2016, 8, 22, 12, tzinfo=pytz.utc)
-        exp = Series([Timedelta('1 days'), pd.NaT])
-        assert_series_equal(s - dt, exp)
-        assert_series_equal(s - Timestamp(dt), exp)
-
-    def test_dt64_series_with_timedelta(self):
-        # scalar timedeltas/np.timedelta64 objects
-        # operate with np.timedelta64 correctly
-        s = Series([Timestamp('20130101 9:01'), Timestamp('20130101 9:02')])
-
-        result = s + np.timedelta64(1, 's')
-        result2 = np.timedelta64(1, 's') + s
-        expected = Series([Timestamp('20130101 9:01:01'),
-                           Timestamp('20130101 9:02:01')])
-        assert_series_equal(result, expected)
-        assert_series_equal(result2, expected)
-
-        result = s + np.timedelta64(5, 'ms')
-        result2 = np.timedelta64(5, 'ms') + s
-        expected = Series([Timestamp('20130101 9:01:00.005'),
-                           Timestamp('20130101 9:02:00.005')])
-        assert_series_equal(result, expected)
-        assert_series_equal(result2, expected)
-
-    def test_dt64_series_add_tick_DateOffset(self):
-        # GH 4532
-        # operate with pd.offsets
-        ser = Series([Timestamp('20130101 9:01'), Timestamp('20130101 9:02')])
-        expected = Series([Timestamp('20130101 9:01:05'),
-                           Timestamp('20130101 9:02:05')])
-
-        result = ser + pd.offsets.Second(5)
-        assert_series_equal(result, expected)
-
-        result2 = pd.offsets.Second(5) + ser
-        assert_series_equal(result2, expected)
-
-    def test_dt64_series_sub_tick_DateOffset(self):
-        # GH 4532
-        # operate with pd.offsets
-        ser = Series([Timestamp('20130101 9:01'), Timestamp('20130101 9:02')])
-        expected = Series([Timestamp('20130101 9:00:55'),
-                           Timestamp('20130101 9:01:55')])
-
-        result = ser - pd.offsets.Second(5)
-        assert_series_equal(result, expected)
-
-        result2 = -pd.offsets.Second(5) + ser
-        assert_series_equal(result2, expected)
-
-        with pytest.raises(TypeError):
-            pd.offsets.Second(5) - ser
-
-    @pytest.mark.parametrize('cls_name', ['Day', 'Hour', 'Minute', 'Second',
-                                          'Milli', 'Micro', 'Nano'])
-    def test_dt64_series_with_tick_DateOffset_smoke(self, cls_name):
-        # GH 4532
-        # smoke tests for valid DateOffsets
-        ser = Series([Timestamp('20130101 9:01'), Timestamp('20130101 9:02')])
-
-        offset_cls = getattr(pd.offsets, cls_name)
-        ser + offset_cls(5)
-        offset_cls(5) + ser
-
-    def test_dt64_series_add_mixed_tick_DateOffset(self):
-        # GH 4532
-        # operate with pd.offsets
-        s = Series([Timestamp('20130101 9:01'), Timestamp('20130101 9:02')])
-
-        result = s + pd.offsets.Milli(5)
-        result2 = pd.offsets.Milli(5) + s
-        expected = Series([Timestamp('20130101 9:01:00.005'),
-                           Timestamp('20130101 9:02:00.005')])
-        assert_series_equal(result, expected)
-        assert_series_equal(result2, expected)
-
-        result = s + pd.offsets.Minute(5) + pd.offsets.Milli(5)
-        expected = Series([Timestamp('20130101 9:06:00.005'),
-                           Timestamp('20130101 9:07:00.005')])
-        assert_series_equal(result, expected)
-
-    def test_dt64_series_sub_NaT(self):
-        # GH#18808
-        dti = pd.DatetimeIndex([pd.NaT, pd.Timestamp('19900315')])
-        ser = pd.Series(dti)
-        res = ser - pd.NaT
-        expected = pd.Series([pd.NaT, pd.NaT], dtype='timedelta64[ns]')
-        tm.assert_series_equal(res, expected)
-
-        dti_tz = dti.tz_localize('Asia/Tokyo')
-        ser_tz = pd.Series(dti_tz)
-        res = ser_tz - pd.NaT
-        expected = pd.Series([pd.NaT, pd.NaT], dtype='timedelta64[ns]')
-        tm.assert_series_equal(res, expected)
-
-    def test_datetime64_ops_nat(self):
-        # GH 11349
-        datetime_series = Series([NaT, Timestamp('19900315')])
-        nat_series_dtype_timestamp = Series([NaT, NaT], dtype='datetime64[ns]')
-        single_nat_dtype_datetime = Series([NaT], dtype='datetime64[ns]')
-
-        # subtraction
-        assert_series_equal(-NaT + datetime_series, nat_series_dtype_timestamp)
-        with pytest.raises(TypeError):
-            -single_nat_dtype_datetime + datetime_series
-
-        assert_series_equal(-NaT + nat_series_dtype_timestamp,
-                            nat_series_dtype_timestamp)
-        with pytest.raises(TypeError):
-            -single_nat_dtype_datetime + nat_series_dtype_timestamp
-
-        # addition
-        assert_series_equal(nat_series_dtype_timestamp + NaT,
-                            nat_series_dtype_timestamp)
-        assert_series_equal(NaT + nat_series_dtype_timestamp,
-                            nat_series_dtype_timestamp)
-
-        assert_series_equal(nat_series_dtype_timestamp + NaT,
-                            nat_series_dtype_timestamp)
-        assert_series_equal(NaT + nat_series_dtype_timestamp,
-                            nat_series_dtype_timestamp)
-
-    @pytest.mark.parametrize('dt64_series', [
-        Series([Timestamp('19900315'), Timestamp('19900315')]),
-        Series([NaT, Timestamp('19900315')]),
-        Series([NaT, NaT], dtype='datetime64[ns]')])
-    @pytest.mark.parametrize('one', [1, 1.0, np.array(1)])
-    def test_dt64_mul_div_numeric_invalid(self, one, dt64_series):
-        # multiplication
-        with pytest.raises(TypeError):
-            dt64_series * one
-        with pytest.raises(TypeError):
-            one * dt64_series
-
-        # division
-        with pytest.raises(TypeError):
-            dt64_series / one
-        with pytest.raises(TypeError):
-            one / dt64_series
-
-    def test_dt64_series_arith_overflow(self):
-        # GH#12534, fixed by #19024
-        dt = pd.Timestamp('1700-01-31')
-        td = pd.Timedelta('20000 Days')
-        dti = pd.date_range('1949-09-30', freq='100Y', periods=4)
-        ser = pd.Series(dti)
-        with pytest.raises(OverflowError):
-            ser - dt
-        with pytest.raises(OverflowError):
-            dt - ser
-        with pytest.raises(OverflowError):
-            ser + td
-        with pytest.raises(OverflowError):
-            td + ser
-
-        ser.iloc[-1] = pd.NaT
-        expected = pd.Series(['2004-10-03', '2104-10-04', '2204-10-04', 'NaT'],
-                             dtype='datetime64[ns]')
-        res = ser + td
-        tm.assert_series_equal(res, expected)
-        res = td + ser
-        tm.assert_series_equal(res, expected)
-
-        ser.iloc[1:] = pd.NaT
-        expected = pd.Series(['91279 Days', 'NaT', 'NaT', 'NaT'],
-                             dtype='timedelta64[ns]')
-        res = ser - dt
-        tm.assert_series_equal(res, expected)
-        res = dt - ser
-        tm.assert_series_equal(res, -expected)
-
-    @pytest.mark.parametrize('op', ['__add__', '__radd__',
-                                    '__sub__', '__rsub__'])
-    @pytest.mark.parametrize('tz', [None, 'Asia/Tokyo'])
-    def test_dt64_series_add_intlike(self, tz, op):
-        # GH#19123
-        dti = pd.DatetimeIndex(['2016-01-02', '2016-02-03', 'NaT'], tz=tz)
-        ser = Series(dti)
-
-        other = Series([20, 30, 40], dtype='uint8')
-
-        pytest.raises(TypeError, getattr(ser, op), 1)
-
-        pytest.raises(TypeError, getattr(ser, op), other)
-
-        pytest.raises(TypeError, getattr(ser, op), other.values)
-
-        pytest.raises(TypeError, getattr(ser, op), pd.Index(other))
 
 
 class TestSeriesOperators(TestData):
